Refactor Barcode Scanner Module
- Removed NutritionEditorSheet and replaced it with NutritionEditorView for improved structure and functionality.
- Introduced OpenFoodFactsClient to handle API requests for product data from OpenFoodFacts.
- Added barcodeScannerOnlyCarbs setting in BolusCalculatorStateModel to allow users to restrict logging to carbohydrates only.
- Updated BolusCalculatorConfigRootView to include a new setting for only allowing carbs from the barcode scanner.
- Changed AppShortcuts to use OpenBarcodeScannerIntent for better integration with Apple Shortcuts.
- Created OpenBarcodeScannerIntent to facilitate opening the barcode scanner via Shortcuts.
